Moved content version edit dialog into separate file
